requirements.txt
numpy
pandas
scikit-learn
py4j
jpype1
matplotlib
emoji
zemberek-python